+++ CHANGELOG.md
@@ -10,6 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -
 
+## [0.3.0] - 2023-07-11
+
+### Chanaged in 0.3.0
+
+- Modified `github.com/senzing/go-messaging/parser`
+- Modified message `details` format
+
 ## [0.2.2] - 2023-06-16
 
 ### Added to 0.2.2